agents/requirements.txt
aiofiles
aiogram
aiosqlite
arrow
python-dotenv
httpx
mastodon.py
quart
